backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
requests
feedparser
datasketch
scikit-learn
sentence-transformers
google-generativeai
python-dotenv
pydantic
